move QubesOS strategy, callback and connection plugins to a collection
Keep a symlink in old locations to avoid breaking changes
